Changed Id to be double, since int couldn't contain the google-id.

+++ SplitListWebApi/ApiFormat/IDTO.cs
@@ -2,7 +2,7 @@
 {
     public interface IDTO
     {
-        int Id { get; set; }
+        double Id { get; set; }
         string Name { get; set; }
     }
 }
