+++ SplitListWebApi/ApiFormat/ShadowTables/PantryItem.cs
@@ -5,9 +5,9 @@ namespace ApiFormat.ShadowTables
 {
     public class PantryItem
     {
-        public int PantryID { get; set; }
+        public double PantryID { get; set; }
         public PantryModel Pantry { get; set; }
-        public int ItemID { get; set; }
+        public double ItemID { get; set; }
         public ItemModel Item { get; set; }
         public int Amount { get; set; }
     }
