+++ SplitListWebApi/ApiFormat/ShadowTables/ShoppingListItem.cs
@@ -5,10 +5,10 @@ namespace ApiFormat.ShadowTables
 {
     public class ShoppingListItem
     {
-        public int ShoppingListID { get; set; }
+        public double ShoppingListID { get; set; }
         public ShoppingListModel ShoppingList { get; set; }
 
-        public int ItemID { get; set; }
+        public double ItemID { get; set; }
         public ItemModel Item { get; set; }
 
         public int Amount { get; set; }
