+++ SplitListWebApi/ApiFormat/User/UserModel.cs
@@ -6,14 +6,15 @@ using Microsoft.AspNetCore.Identity;
 
 namespace ApiFormat.User
 {
-    public class UserModel : IdentityUser<int>, IModel
+    public class UserModel : IdentityUser<double>, IModel
     {
+
         [Key, DatabaseGenerated(DatabaseGeneratedOption.None)]
-        public override int Id { get; set; }
+        public override double Id { get; set; }
         public List<UserGroup> UserGroups { get; set; }
         public string Name { get; set; }
     }
-    public class ApplicationRole : IdentityRole<int>
+    public class ApplicationRole : IdentityRole<double>
     {
 
     }
